[fix] rename getTodos to fetchTodos, rename user-slice and user-actions to auth-slice and auth-actions for upcoming admin/users feature
