+++ src/App.tsx
@@ -11,7 +11,7 @@ import AuthLayout from "./pages/AuthLayout/AuthLayout";
 import AuthPage from "./pages/AuthPage";
 import SignupPage from "./pages/SignupPage";
 import ProtectedRoute from "./pages/ProtectedRoute";
-import { silentRefreshThunk } from "./store/user/user-actions";
+import { silentRefreshThunk } from "./store/auth/auth-actions";
 
 const router = createBrowserRouter([
   {
