+++ src/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import todosReducer from "./todos/todos-slice";
 import uiReducer from "./ui/ui-slice";
-import userReducer from "./user/user-slice";
+import userReducer from "./auth/auth-slice";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 
 export const store = configureStore({
